+++ backend/src/Entity/Character.php
@@ -42,7 +42,7 @@ class Character implements \JsonSerializable
      * @ORM\Column(type="string", length=255)
      * @var string
      */
-    private $name;
+    private $name = '';
 
     /**
      * @OA\Property()