+++ frontend/src/pages/Home.vue
@@ -231,12 +231,6 @@ export default {
             if (! this.player) {
                 return;
             }
-            const vm = this;
-            this.player.characters.forEach(function(character) {
-                if (character.lastUpdate === null) {
-                    vm.update(character.id);
-                }
-            });
             this.checkManaged();
             this.checkDeactivated();
         },